# common: optimize OpTracker