--- a/src/common/TrackedOp.cc
+++ b/src/common/TrackedOp.cc
@@ -189,7 +189,6 @@ OpTracker::OpTracker(CephContext *cct_, bool tracking, uint32_t num_shards):
   log_threshold(0),
   tracking_enabled(tracking),
   cct(cct_) {
-  assert(ISP2(num_shards));
 }
 
 OpTracker::~OpTracker() {
@@ -284,8 +283,7 @@ bool OpTracker::register_inflight_op(TrackedOp *i)
     return false;
 
   uint64_t current_seq = ++seq;
-  uint32_t shard_index = current_seq & (sharded_in_flight_list.size() - 1);
-  ShardedTrackingData& sdata = sharded_in_flight_list[shard_index];
+  ShardedTrackingData& sdata = sharded_in_flight_list.get_shard(current_seq);
   {
     Mutex::Locker locker(sdata.ops_in_flight_lock_sharded);
     sdata.ops_in_flight_sharded.push_back(*i);
@@ -300,8 +298,7 @@ void OpTracker::unregister_inflight_op(TrackedOpRef&& i)
   // caller checks;
   assert(ri->state);
 
-  uint32_t shard_index = ri->seq & (sharded_in_flight_list.size() - 1);
-  ShardedTrackingData& sdata = sharded_in_flight_list[shard_index];
+  ShardedTrackingData& sdata = sharded_in_flight_list.get_shard(ri->seq);
   {
     Mutex::Locker locker(sdata.ops_in_flight_lock_sharded);
     auto p = sdata.ops_in_flight_sharded.iterator_to(*ri);