--- a/src/common/TrackedOp.h
+++ b/src/common/TrackedOp.h
@@ -34,9 +34,9 @@ class OpHistoryServiceThread : public Thread
     utime_t time;
     TrackedOpRef op;
 
-    queue_item_t(utime_t time, TrackedOp* op)
+    queue_item_t(utime_t time, TrackedOpRef&& op)
       : time(std::move(time)),
-        op(TrackedOpRef(op)) {
+        op(std::move(op)) {
     }
   };
   typedef boost::intrusive::list<queue_item_t> queue_t;
@@ -53,8 +53,8 @@ public:
   ~OpHistoryServiceThread();
 
   void break_thread();
-  void insert_op(const utime_t& now, TrackedOp* op) {
-    auto item = new queue_item_t(now, op);
+  void insert_op(const utime_t& now, TrackedOpRef&& op) {
+    auto item = new queue_item_t(now, std::move(op));
     queue_spinlock.lock();
     _external_queue.push_back(*item);
     queue_spinlock.unlock();
@@ -90,12 +90,12 @@ public:
     assert(duration.empty());
     assert(slow_op.empty());
   }
-  void insert(const utime_t& now, TrackedOp* op)
+  void insert(const utime_t& now, TrackedOpRef&& op)
   {
     if (shutdown)
       return;
 
-    opsvc.insert_op(now, op);
+    opsvc.insert_op(now, std::move(op));
   }
 
   void _insert_delayed(const utime_t& now, TrackedOpRef op);
@@ -322,7 +322,7 @@ public:
   bool dump_historic_ops(Formatter *f, bool by_duration = false, set<string> filters = {""});
   bool dump_historic_slow_ops(Formatter *f, set<string> filters = {""});
   bool register_inflight_op(TrackedOp *i);
-  void unregister_inflight_op(TrackedOp *i);
+  void unregister_inflight_op(TrackedOpRef&& i);
 
   void get_age_ms_histogram(pow2_hist_t *h);
 
@@ -403,25 +403,30 @@ inline void TrackedOp::tracking_start() {
 }
 
 inline void TrackedOp::put() {
-  if (--nref == 0) {
+  if (nref.load() == 1) {
     switch (state.load()) {
     case STATE_UNTRACKED:
+      --nref;
       _unregistered();
       delete this;
       break;
 
     case STATE_LIVE:
       mark_event("done");
-      tracker->unregister_inflight_op(this);
+      tracker->unregister_inflight_op(
+	TrackedOpRef(this, false /* = add_ref */));
       break;
 
     case STATE_HISTORY:
+      --nref;
       delete this;
       break;
 
     default:
       ceph_abort();
     }
+  } else {
+    --nref;
   }
 }