… src/common/align.h
… src/common/ceph_context.cc
… src/common/containers.h
… src/include/intarith.h
… src/include/spinlock.h
… src/test/simple_spin.cc