Summary
=======
Last week the `OpHistory` overhead has been wisely removed from the main path. I observe several % less `futex` syscalls per read on the fresh `master` :-). This pull request tries to deal with `OpTracker`, an entity that sits above the `OpHistory` and remains a hot spot.

In the matter of `cycles:pp` before the PR #20540 (quite old `master`):
```
         - 3,16% PGOpItem::~PGOpItem
            - 2,91% OpTracker::unregister_inflight_op
                 1,25% OpHistory::insert
                 0,58% OpRequest::~OpRequest
```

After PR #20540 but before this one:
```
         - 2,34% PGOpItem::~PGOpItem
            - TrackedOp::put
                 2,10% OpTracker::unregister_inflight_op
```

Notable changes
============
* Eradicate the `RWLock`-typed `OpTracker::lock` that is being used on the main path but **never gets locked exclusively**. This suggests it's virtually a no-op associated with non-zero cost, higher in comparison to plain mutex. For details please refer to [the discussion](http://permalink.gmane.org/gmane.comp.lib.boost.devel/211180) regarding real functionality of `std::shared_mutex` and the lack of its standardization in C++11. The extra ref counting we do by default in `RWLock` rises the cost further.
* Change the way how `tp_osd_tp` threads synchronize with the new `OpHistory` management thread. The optimizing started from improving the `ceph::spinlock` implementation but ultimately moved to plain `std::mutex` treated as a spinlock-with-backoff-ability. The main idea is to prevent situations when e.g. preemption of single worker thread makes all others (up to 16 - 1 by default) burning CPU and starving everything else. That's the purpose of introducing `ceph::spin::adapt_guard`. It might be useful in other places as the `pthread_mutex_lock` from glibc  effectively `try_lock` once and then fallback to kernel (applies to the `PTHREAD_MUTEX_NORMAL`).
* Improve `ceph::spinlock`. Although ultimately dropped from `OpHistoryServiceThread`, optimization in this area still could be beneficial for other users (e.g. `AsyncMessanger`, `CephContext`).
The current implementation uses `lock cmpxchg` in very tight loop which isn't the way to go, especially on SMT systems. Please refer to [Long Duration Spin-wait Loops on Hyper-Threading Technology Enabled Intel Processors](https://software.intel.com/en-us/articles/long-duration-spin-wait-loops-on-hyper-threading-technology-enabled-intel-processors/) and [Benefitting Power and Performance Sleep Loops](https://software.intel.com/en-us/articles/benefitting-power-and-performance-sleep-loops). Although being Intel's papers, I believe they apply to recent AMD's CPUs as well. Additionally, relaxation of memory ordering in the lock phase *might* (untested!) benefit machines with weaker memory model than x86.
* Remove the nasty pattern of `division-missed load-missed load` during shard access with `ceph::tiny_vector` and through making the assumption that number of `OpTracker`'s shards is p2. The further idea is to combine `ceph::tiny_vector` with `ceph::math::p2_t` to form `ceph::containers::shard_vector` which would also take care of enforcing cache line alignment.
* Kill the ref counting on op deregistration. On x86 locked instructions are necessary for the sake of atomicity. They act just like a memory barrier.

Performance comparison
==================

All tests have been conducted on a single, physical *incerta* node in 5 rounds. The result for each round is sum of **bandwidth** 4 separated FIO-librbd clients get from the cluster.


Reference point: 8e8c5e2a3849c90f4f16892e1150d22de8f4de34.
-----------------------------------------------------------------------------------
```
randread,  4 KiB:  355585,  348867,  353343,  356998,  347209
randread, 64 KiB: 3176554, 3183690, 3189953, 3181545, 3046893
```

After the changes: ee8ed2bc48c6bc8dbc9cd70fb1c515fd6642429a.
---------------------------------------------------------------------------------------
```
randread,  4 KiB:  371652,  377484,  371640,  371964,  372908
randread, 64 KiB: 3262052, 3236116, 3256350, 3244601, 3229833

```

Profiling in the domain of `cycles:pp`
==========================
Everything happened on my development machine (Intel Core i7-6820HQ, Samsung SM961 NVMe).

### Before PR #20540
```
   - 54,36% ShardedThreadPool::shardedthreadpool_worker
      - 54,19% OSD::ShardedOpWQ::_process
         + 45,25% PGOpItem::run
         - 3,16% PGOpItem::~PGOpItem
            - 2,91% OpTracker::unregister_inflight_op
                 1,25% OpHistory::insert
                 0,58% OpRequest::~OpRequest
         + 1,48% Cond::Wait
         + 1,40% PG::lock
         + 0,87% WeightedPriorityQueue<OpQueueItem, unsigned long>::dequeue
           0,52% Mutex::Unlock
```


### After PR #20540 but before this one
```
-   52,51%     0,13%  tp_osd_tp        libceph-common.so.0   [.] ShardedThreadPool::shardedthreadpool_worker
   - 52,38% ShardedThreadPool::shardedthreadpool_worker
      - 52,23% OSD::ShardedOpWQ::_process
         + 43,90% PGOpItem::run
         - 2,34% PGOpItem::~PGOpItem
            - TrackedOp::put
                 2,10% OpTracker::unregister_inflight_op
         + 1,57% Cond::Wait
         + 1,47% PG::lock
         + 0,76% WeightedPriorityQueue<OpQueueItem, unsigned long>::dequeue
           0,55% Mutex::Unlock
```


### Killing the `OpTracker::lock` has exposed the spinlock contention

This varies greatly. I observed values between 5-11%.
```
-   56,13%     0,11%  tp_osd_tp        libceph-common.so.0   [.] ShardedThreadPool::shardedthreadpool_worker
   - 56,02% ShardedThreadPool::shardedthreadpool_worker
      - 55,89% OSD::ShardedOpWQ::_process
         + 40,99% PGOpItem::run
         - 9,50% PGOpItem::~PGOpItem
            - TrackedOp::put
                 9,29% OpTracker::unregister_inflight_op
         + 1,50% Cond::Wait
         + 1,32% PG::lock
         + 0,67% WeightedPriorityQueue<OpQueueItem, unsigned long>::dequeue
```


### At the moment
Hopefully not finished yet -- there is the `register_inflight_op` part to deal with. :-)
```
   - 51,51% ShardedThreadPool::shardedthreadpool_worker
      - 51,34% OSD::ShardedOpWQ::_process
         + 44,12% PGOpItem::run
         + 1,64% Cond::Wait
         + 1,52% PG::lock
         - 1,12% PGOpItem::~PGOpItem
            - TrackedOp::put
               - 0,87% OpTracker::unregister_inflight_op
                    0,54% OpRequest::_unregistered
         + 0,75% WeightedPriorityQueue<OpQueueItem, unsigned long>::dequeue
           0,55% Mutex::Unlock
```

**UPDATE**: I plan to continue with `register_inflight_op` after PR #19973 gets merged. The idea is to establish a binding between shards of `OpTracker`  and `ShardedOpWQ` with `spg_t` as a selector. At least for the OSD use case, this would make the `std::atomic<int64_t>`-typed `OpTracker::seq` unnecessary.