@branch-predictor 2018-04-13
tl;dr - too much changes, too small performance version. 

longer tl;dr: the perf difference is there, but only when running on overloaded, underpowered hardware.

Long version:

So in a nutshell, here's what changed:
* replaced old spinlock with new one that uses PAUSE instruction internally 
* disabled refcounting on RWLock
* got rid of memory allocation under spinlock in insert_op
* replaced vector with different vector
* got rid of two divisions 

and that gave us 2-4% bandwidth increase when running on laptop and single consumer SSD. 

@rzarzynski According to your own measurements, you noticed between 2 to 4% of *bandwidth* increase, which quite small considering the amount of code changes, To add insult to the injury, you measured it on a *mobile* cpu with 4 cores  (8 threads), meaning all these things are way more inflated than in reality (nobody is going to run 3+ production OSD cluster along with 4 clients on laptop). I like the spinlock re-implementation - also, RWLock and memory allocation (which I left out to not pollute my PR any further) changes are always welcome, but the rest isn't exactly profitable. 

I did a quick check of your patch on vstart cluster with lockdep disabled and with bluestore, running on Xeon E5-2687W v3, and I see no impact on performance at all - the load on optracker's service thread is the same (around 4%) and the number of IOPS achieved by three running in parallel 4k ```rados bench read -t 128 --no-verify``` on bluestore remains on the same level (~74K IOPS). Op latency varies and there's no clear winner. 

To reiterate - I believe in your results, they look legitimate considering the machine and load, but this doesn't exactly justify the code churn and don't assert reasonable performance improvement in actual environment. But if the current objective is to optimize for underpowered machines, I still have Atom 330 machine - I'm sure that kind of CPU will see such microoptimizations with open arms.