@branch-predictor 2018-04-18
On 18-04-18 12:48 AM, Radoslaw Zarzynski wrote:
>     @rzarzynski <https://github.com/rzarzynski> According to your own
>     measurements, you noticed between 2 to 4% of bandwidth increase, which
>     quite small considering the amount of code changes,
> 
> Could you please elaborate a bit on deriving the numbers above? My 
> calculations give:
> 
>   * 4 KiB randreads:
>       o min difference: |371640. / 356998 ~=|*|1.0410|*,
>       o max difference: |377484. / 347209 ~=|*|1.0872|*,
>       o on average: |(371652 + 377484 + 371640 + 371964 + 372908.) / (355585
>         + 348867 + 353343 + 356998 + 347209.) ~=|*|1.0588|*.
>   * 64 KiB randreads:
>       o min difference: |3229833. / 3189953 ~=|*|1.0125|*,
>       o max difference: |3262052. / 3046893 ~=|*|1.0706|*,
>       o on average: |(3262052 + 3236116 + 3256350 + 3244601 + 3229833.) /
>         (3176554 + 3183690 + 3189953 + 3181545 + 3046893) ~=|*|1.0285|*.

Doesn't really matter, as you didn't specify the units, so it's not obvious 
whether that "3229833" is bandwidth in bytes, megabytes or bits. If that's 
bytes, then your measurements are even stronger proof that environment was 
too underpowered to gather reliable results - on mentioned Xeon I was able 
to go beyond GB per second in 64k rand reads.

>     I like the spinlock re-implementation - also, RWLock and memory
>     allocation (which I left out to not pollute my PR any further) changes
>     are always welcome, but the rest isn't exactly profitable.
> 
> Why do think the rest isn't (exactly) profitable? For an out-of-order 
> microarchitecture it's quite usual that bottlenecks do overlap and -- to see 
> the benefit fully -- all of them need to be optimized. For instance, the 
> |div| in |unregister_inflight_op| popped up after a couple of commits 
> (cycles; profiled with the Precision Event-Based Sampling):
> 
> |OpTracker::unregister_inflight_op /work/ceph-1/build/bin/ceph-osd Percent│ 
> mov %rsi,%rbx 0,56 │ mov %rdi,%rbp │ _ZN5Mutex6LockerC4ERS_(): │ class 
> Locker { │ Mutex &mutex; │ │ public: │ explicit Locker(Mutex& m) : mutex(m) 
> { │ mutex.Lock(); │ xor %esi,%esi │ 
> _ZN9OpTracker22unregister_inflight_opEP9TrackedOp(): 34,83 │ divq 0x8(%rdi) |

Because at line 314 you still have:

```const utime_t now = ceph_clock_now();```

which is asking glibc to return current real time - which by itself is slow, 
and there's also object construction during which two __u32's are copied.
I even measured the worst-case timings for you and it turns out that my Xeon 
is capable of calculating ~380 000 000 integer modulo operations per second 
and execute "clock_gettime(CLOCK_REALTIME, ...)" alone ~39 000 000 times per 
second. That makes it impossible for division you optimized to take 35% of 
cpu time because it's not ~2 times faster than clock_gettime, but about 10 
times faster, and certainly even more than ceph_clock_now() - which makes 
sense when you consider that it's a single CPU instruction vs function calls 
[branching!] and a few tens of instructions. Sure, the (original) code that 
follows is stalled by (that is, can't execute in parallel/separate pipeline 
because it waits for result of) modulo operation, but it doesn't really 
matter because both original and your changed code will be badly stalled 
anyway by call to ceph_clock_now().
Now, switching to CLOCK_REALTIME_COARSE will yield 4x reduction in cost of 
clock_gettime -  that can execute ~170 000 000 times a second, which is half 
of division. Bit operators are very fast, in my short testing it can do 
around 2 000 000 000 bit ands per second. Not sure whether it's valid number 
but let's pick it anyway. Now let's do the math:

```
* modulo:          1/380 000 000 s = 0.000000003
* clock_gettime:    1/39 000 000 s = 0.000000026
```

That gives us 0.000000029s per iteration of unregister_inflight_op.
What changes after optimizing division?

```
* bit and:       1/2 000 000 000 s = 0.0000000005
* clock_gettime:    1/39 000 000 s = 0.000000026
```
That gives is 0,0000000265s which is around 9% improvement. Even if you 
assume that it can execute 20 000 000 000 bit ands per second, you still are 
in 11% improvement area. Unfortunately, even that's not the case here, 
because operator "%" for p2_t is defined as follows:

```
  friend value_type operator%(const value_type& l,
                              const p2_t<value_type>& r) {
     return l & (r.value - 1);
  }
```

which means its bitwise "and" AND subtraction, and the result of bitwise 
"and" depends on the result of subtraction (no ability to execute both 
calculations in two separate pipelines, even if there are two free pipelines 
at the moment). My xeon is capable of doing 1 600 000 000 of these pairs per 
second, but let's skip that detail because it's even more irrelevant. What 
if we replace CLOCK_REALTIME with CLOCK_REALTIME_COARSE?

```
* modulo:                 1/380 000 000 s = 0.000000003
* CLOCK_REALTIME_COARSE:  1/170 000 000 s = 0.000000006
```

Total time is 0.000000009, which is over 3x times faster than it used to be.
These numbers are of course very rough and not to be taken seriously, but 
hopefully picture the problem well enough.
If you take that into consideration, you'll realize that 9/11% perf increase 
coming from removal of int division isn't exactly big win and surely is 
confusing, because suddenly the int modulo operator isn't an int modulo 
operator but bitwise "and", because that's different var type that's defined 
in extra code that needs to be maintained. You could achieve the same result 
by enforcing num_optracker_shards to be a power of two and using bitwise 
"and" explicitly. It's much less code churn with even better result - no 
subtraction!

> Even if it's responsible for 0,15% of cycles, we had 2 of them in the 
> affected area. Moreover, the share is measured in the scope of whole 
> |ceph-osd| process. For single |tp_osd_tp| it's higher.

It's only an artifact of sampling profiler working on function level, 
hitting similar spot all the time. That's not 35% of time spent, but 35% of 
samples.

>     I did a quick check of your patch on vstart cluster with lockdep
>     disabled and with bluestore, running on Xeon E5-2687W v3, and I see no
>     impact on performance at all - the load on optracker's service thread is
>     the same (around 4%) and the number of IOPS achieved by three running in
>     parallel 4k |rados bench read -t 128 --no-verify| on bluestore remains
>     on the same level (~74K IOPS).
> 
> Do you have a chance to check the utilization of |tp_osd_tp| and 
> |msgr-worker-*|? I wouldn't expect a difference on |OpHistorySvc| as its 
> internal processing isn't affected by the PR. Even better would be to employ 
> CBT.

I'm afraid I don't see any meaningful difference either.

> What makes me afraid is that there are many things, both in software and 
> hardware, that can affect the results. This starts from benchmarks being 
> used, goes through deployment/configuration, incorporates the number of CPU 
> sockets and -- for Haswells -- even a microcode version (TSX-related 
> hardware bugs), finally ending on operating system level. A cluster might be 
> just bottlenecked earlier. There are so many degrees-of-freedom that I get a 
> headache and direct towards selection of a single reference point. I'm 
> unable to say how well /incerta/ matches a /typical/ high-end configuration 
> in the field (any feedback is more than welcome!). However, I treat it as a 
> scripted blackbox giving useful /thumb-up/ and /thumb-down/ signals.

And that's why I'm insisting on focusing more on performance bottlenecks 
that are more obvious in their nature and more beneficial for everyone 
*ESPECIALLY* if that optimization involves writing code that's not 
immediately obvious. You might write some super-fast, super-optimal code on 
x86, but on ARM or m68k it might even not run correctly. You might write a 
code that's fast for limited set of cases. You might write a code that 
depends on compiler optimization. You might write a code that's fast on 
particular iteration of cpu's internal architecture and on another it takes 
a deep dive. It takes *YEARS* to fully understand what code to expect from 
compilers, how current processors operate and how they interact with 
hardware, and then you just give up and optimize on algorithm level, taking 
into consideration limited set of basic rules that are easy to understand 
without going too deep into how processors operate.
>     But if the current objective is to optimize for underpowered machines, I
>     still have Atom 330 machine - I'm sure that kind of CPU will see such
>     microoptimizations with open arms.
> 
> I consider these words as unnecessary.

Why? If Ceph can run on ARM microcontrollers built into HDD logic board, 
what's the problem with running it on low-power x86 CPU?